requirements.txt
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.7.0
attrs==24.3.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
cattrs==23.2.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
ell-ai==0.0.15
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.7
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.18.4
numpy==2.2.0
openai==1.58.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==5.9.8
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.37.2
streamlit==1.41.1
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
watchdog==6.0.0
watchfiles==1.0.3
websockets==14.1
